+++ tests/test_table.py
@@ -1,9 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf8 -*-
 
-from decimal import Decimal
-import warnings
-
 try:
     import unittest2 as unittest
 except ImportError:
@@ -80,44 +77,3 @@ class TestTable(unittest.TestCase):
 
         self.assertEqual(len(new_table.rows), 1)
         self.assertSequenceEqual(new_table.columns['one'], (200,))
-
-    def test_pearson_correlation(self):
-        rows = (
-            (-1, 0, 'a'),
-            (0, 0, 'b'),
-            (1, 3, 'c')
-        )
-
-        table = agate.Table(rows, self.column_names, self.column_types)
-
-        self.assertEqual(table.pearson_correlation('one', 'one'), Decimal('1'))
-        self.assertAlmostEqual(table.pearson_correlation('one', 'two'), Decimal('3').sqrt() * Decimal('0.5'))
-
-    def test_pearson_correlation_nulls(self):
-        rows = (
-            (-1, 0, 'a'),
-            (0, 0, 'b'),
-            (1, None, 'c')
-        )
-
-        table = agate.Table(rows, self.column_names, self.column_types)
-
-        warnings.simplefilter('error')
-
-        with self.assertRaises(agate.NullCalculationWarning):
-            table.pearson_correlation('one', 'two')
-
-        warnings.simplefilter('ignore')
-
-        self.assertEqual(table.pearson_correlation('one', 'two'), 0)
-
-    def test_pearson_correlation_zero(self):
-        rows = (
-            (-1, 3, 'a'),
-            (0, 3, 'b'),
-            (1, 3, 'c')
-        )
-
-        table = agate.Table(rows, self.column_names, self.column_types)
-
-        self.assertEqual(table.pearson_correlation('one', 'two'), Decimal('0'))