+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 0.3.0
 -----
 
+* Convert :class:`.PearsonCorrelation` to an aggregation.
 * Update required version of agate to 1.1.0.
 * Removed Python 2.6 support.